Add setting to filter test executables
Currently, the adapter treats all executables as test executables. This is very inconvenient when the solution includes other executables, which get run on every test discovery round.

Implemented a setting for filtering the tests. Since this is the first setting for the adapter, had to initialize the infrastructure for reading settings from the runsettings file.