Fixed: Test input and output files stored in VS directory
The fix to the problem with directories with spaces transitioned
to moving relative path names for the test input and output files.
Sadly, the current working directory of the test process
is some wandom Visual Studio directory, which caused problems
with access rights sometimes, and is not a great
place for them anyway.

Fixed by explicitly setting the working directory based on
the test executable.