Use DTE to get the path to the solution
VS2015 does not provide the solution path in the ConfigurationInfo when it calls AddRunSettings in the test Discovery phase. This made it necessary to fetch the solution path from the global DTE object instead.

Added the assembly that defines the DTE interface as a NuGet package. Visual Studio provides it when installed, but a nuget package should be more reliable in CI environments.